Code clean up: Put Components into a serparate file to match others and reduce Controler file code
